# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: React Sheets Workbook API
  version: 2026-08-24
servers:
  - url: /
paths:
...
  /api/auth/bootstrap:
    post:
      operationId: bootstrapLocalAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
      responses:
        '200':
          description: Established administrator session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
  /api/auth/login:
    post:
      operationId: loginLocalUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Established user session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
...
components:
...
  schemas:
...
    AuthSessionResponse:
      type: object
      required: [authenticated, subject, displayName, admin, bootstrapRequired, csrfToken]
      properties:
        authenticated: { type: boolean }
        subject: { type: [string, 'null'] }
        displayName: { type: [string, 'null'] }
        admin: { type: boolean }
        bootstrapRequired: { type: boolean }
        csrfToken: { type: string }
    BootstrapRequest:
      type: object
      required: [token, username, password, displayName]
      properties:
        token: { type: string, minLength: 1, maxLength: 512 }
        username: { type: string, minLength: 1, maxLength: 200 }
        password: { type: string, minLength: 12, maxLength: 200 }
        displayName: { type: string, minLength: 1, maxLength: 200 }
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, minLength: 1, maxLength: 200 }
        password: { type: string, minLength: 12, maxLength: 200 }
...